+++ posthog/CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Next
 
+- feat: Add ETag support for local evaluation polling to reduce bandwidth ([#350](https://github.com/PostHog/posthog-android/pull/350))
+- **breaking** (internal): `PostHogApi.localEvaluation()` now returns `LocalEvaluationApiResponse` instead of `LocalEvaluationResponse?` ([#350](https://github.com/PostHog/posthog-android/pull/350))
+- **breaking** (internal): `LocalEvaluationApiResponse.isNotModified` renamed to `wasModified` (inverted) ([#350](https://github.com/PostHog/posthog-android/pull/350))
+
 ## 5.2.0 - 2025-12-03
 
 - feat: include `evaluated_at` properties in `$feature_flag_called` events ([#321](https://github.com/PostHog/posthog-android/pull/321))